working on the cleaning function
the cleaning function requires the files in the folder to be of similar name. I moved the files that were not used in the hard code cleaning process to a sub folder.

I thought of a possible issue that may arise in the future. when we find additional datasets to use in the project, the cleaning function will not work for them if they are not of similar name to the first dataset we used.